Put the beginning part of the survey into the survey app and made the homePage app more like a homePage app
