+++ homePage/forms.py
@@ -1,50 +1,2 @@
-from django import forms
 
 
-class HomePageForm(forms.Form):
-    SURVEY_TYPES = [
-        ('rent', 'Rent'),
-        ('buy', 'Buy'),
-    ]
-    survey_type = forms.ChoiceField(
-        choices=SURVEY_TYPES,
-        widget=forms.RadioSelect()
-    )
-
-    streetAddress = forms.CharField(
-        label="Destination",
-        widget=forms.TextInput(
-            attrs={
-                'class': 'form-control',
-                'placeholder': 'Enter in a Destination',
-                'autocomplete': 'off',
-            }),
-        max_length=200,
-    )
-
-    city = forms.CharField(
-        widget=forms.TextInput(
-            attrs={
-                'class': 'form-control',
-                'placeholder': 'Enter the city',
-            }),
-        max_length=200,
-    )
-
-    state = forms.CharField(
-        widget=forms.TextInput(
-            attrs={
-                'class': 'form-control',
-                'placeholder': 'Enter the State',
-            }),
-        max_length=200,
-    )
-
-    zip_code = forms.CharField(
-        widget=forms.TextInput(
-            attrs={
-                'class': 'form-control',
-                'placeholder': 'Enter the Zip Code',
-            }),
-        max_length=200,
-    )
