+++ survey/templates/survey/index.html
@@ -78,12 +78,9 @@
     <!-- Main component for a primary marketing message or call to action -->
     <div class="jumbotron">
         <h1>Find your Perfect Home!</h1>
-        <p>Welcome to Unicorn! :)</p>
+        <p>Please fill out the survey below to find your unicorn :)</p>
         <p>
-            Please log in to save your survey results
-        </p>
-        <p>
-            <a class="btn btn-lg btn-success" href="{% url 'userAuth:login' %}" role="button">Log In</a>
+            <a class="btn btn-lg btn-success" href="#" role="button">Log In</a>
         </p>
     </div>
     <div class="row">
@@ -95,11 +92,44 @@
                     </div>
                 </div>
             {% endif %}
+            <div class="row">
+                <form class="form-horizontal" action="{% url 'survey:index' %}" method="post">
+                    {% csrf_token %}
+                    <div class="col-sm-offset-3 col-sm-6">
+                        <div class="form-group">
+                            {% for radio in form.survey_type %}
+                                {{ radio }}
+                            {% endfor %}
+                        </div>
+                        <div class="form-group">
+                            <div class="streetAddressDiv">
+                                {{ form.streetAddress.errors }}
+                                {{ form.streetAddress.label_tag }}
+                                {{ form.streetAddress }}
+                            </div>
+                            <div class="cityDiv">
+                                {{ form.city.errors }}
+                                {{ form.city.label_tag }}
+                                {{ form.city }}
+                            </div>
+                            <div class="stateDiv">
+                                {{ form.state.errors }}
+                                {{ form.state.label_tag }}
+                                {{ form.state }}
+                            </div>
+                            <div class="zipCodeDiv">
+                                {{ form.zip_code.errors }}
+                                {{ form.zip_code.label_tag }}
+                                {{ form.zip_code }}
+                            </div>
+                        </div>
+                        <div class="form-group">
+                            <button type="submit" id="Submit" class="btn btn-primary">Submit Survey</button>
+                        </div>
+                    </div>
+                </form>
+            </div>
         </div>
-    </div>
-
-    <div>
-        <a class="btn btn-lg btn-primary" href="{% url 'survey:index' %}">Start Survey</a>
     </div>
