+++ survey/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 
 from . import views
 
-app_name = 'userAuth'
+app_name = 'survey'
 urlpatterns = [
     url(r'^$', views.index, name='index'),
 ]
